+++ tests/Feature/Stories/OnlyGroupAdministratorCanUpdateGroupTest.php
@@ -6,10 +6,12 @@ namespace Tests\Feature\Stories;
 
 use App\Group;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class OnlyGroupAdministratorCanUpdateGroupTest extends TestCase
 {
+    use RefreshDatabase;
     /**
      * @test testUserGroupAdministratorCanUpdateGroup
      */
