+++ tests/Feature/Stories/UserCanCreateGroupsTest.php
@@ -6,12 +6,14 @@ namespace Tests\Feature\Stories;
 
 use App\Group;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use Tests\TestCase;
 
 class UserCanCreateGroupsTest extends TestCase
 {
+    use RefreshDatabase;
     /**
      * @test testUserCanStoreGroup
      */
