+++ tests/Feature/Stories/UserCanGetGeofencesPhotosTest.php
@@ -7,10 +7,13 @@ namespace Tests\Feature\Stories;
 use App\Geofence;
 use App\Photo;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class UserCanGetGeofencesPhotosTest extends TestCase
 {
+    use RefreshDatabase;
+
     /**
      * @test testIndexGeofencesPhotosApi
      */
