+++ tests/Feature/Stories/UserCanViewOwnedPhotosTest.php
@@ -6,10 +6,12 @@ namespace Tests\Feature\Stories;
 
 use App\Photo;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class UserCanViewOwnedPhotosTest extends TestCase
 {
+    use RefreshDatabase;
     /**
      * @test testIndexUsersPhotosApi
      */
