+++ tests/Feature/Stories/UserCanViewPetPhotosTest.php
@@ -7,10 +7,12 @@ namespace Tests\Feature\Stories;
 use App\Pet;
 use App\Photo;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class UserCanViewPetPhotosTest extends TestCase
 {
+    use RefreshDatabase;
     /**
      * @test testIndexPetsPhotosApi
      */
