+++ tests/Feature/Stories/UserGroupMemberCanViewOtherGroupMembersTest.php
@@ -6,10 +6,12 @@ namespace Tests\Feature\Stories;
 
 use App\Group;
 use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class UserGroupMemberCanViewOtherGroupMembersTest extends TestCase
 {
+    use RefreshDatabase;
     /**
      * @test testGroupMemberUserCanViewRelatedGroupUsers
      */
